update split SingleProduct
