+++ src/components/PopularProduct/Product.module.css
@@ -73,9 +73,6 @@ svg polyline {
 }
 
 @media screen and (max-width:576px) {
-  .product_img {
-    height: 230px;
-  }
   .swiperButton {
     display: flex;
   }
